requirements.txt
beautifulsoup4>=4.12
faiss-cpu>=1.8
jieba>=0.42
markdown>=3.6
numpy>=1.26
openai>=1.0
pandas>=2.0
ragas>=0.3,<0.5
sentence-transformers>=2.7
